+++ src/windows.rs
@@ -3,28 +3,87 @@
 // Copyright (C) 2023-2024 James Petersen <m@jamespetersen.ca>
 // Licensed under Apache 2.0 OR MIT. See LICENSE-APACHE or LICENSE-MIT
 
-use std::{alloc::{alloc_zeroed, dealloc, Layout}, mem::align_of, ops::Deref, path::PathBuf, ptr::null_mut, string::FromUtf16Error};
+use core::fmt;
+use std::{
+    alloc::{alloc_zeroed, dealloc, Layout},
+    mem::align_of,
+    ops::Deref,
+    path::PathBuf,
+    ptr::null_mut,
+};
 
 use cfg_if::cfg_if;
-use widestring::{error::{ContainsNul, Utf16Error}, U16CStr, U16CString, U16Str};
-use windows::{core::{w, Error as WinError, BSTR, PCWSTR, PWSTR, VARIANT}, Win32::{Foundation::{CloseHandle, LocalFree, CO_E_NOTINITIALIZED, ERROR_INSUFFICIENT_BUFFER, ERROR_NONE_MAPPED, HANDLE, HLOCAL, PSID}, Security::{Authorization::ConvertSidToStringSidW, GetTokenInformation, LookupAccountNameW, TokenUser, SID, SID_NAME_USE, TOKEN_QUERY, TOKEN_USER}, System::{Com::{CoCreateInstance, CoInitializeEx, CoSetProxyBlanket, CoTaskMemFree, CLSCTX_INPROC_SERVER, COINIT_MULTITHREADED, EOAC_NONE, RPC_C_AUTHN_LEVEL_CALL, RPC_C_IMP_LEVEL_IMPERSONATE}, Rpc::{RPC_C_AUTHN_WINNT, RPC_C_AUTHZ_NONE}, Threading::{GetCurrentProcess, OpenProcessToken}, Wmi::{IWbemLocator, IWbemServices, WbemLocator, WBEM_FLAG_CONNECT_USE_MAX_WAIT, WBEM_FLAG_FORWARD_ONLY, WBEM_FLAG_RETURN_IMMEDIATELY, WBEM_INFINITE}}, UI::Shell::{FOLDERID_Profile, SHGetKnownFolderPath, KNOWN_FOLDER_FLAG}}};
+use widestring::{
+    error::{ContainsNul, Utf16Error},
+    U16CStr, U16CString, U16Str,
+};
+use windows::{
+    core::{w, Error as WinError, BSTR, PCWSTR, PWSTR, VARIANT},
+    Win32::{
+        Foundation::{
+            CloseHandle, LocalFree, CO_E_NOTINITIALIZED, ERROR_INSUFFICIENT_BUFFER,
+            ERROR_NONE_MAPPED, HANDLE, HLOCAL, PSID,
+        },
+        Security::{
+            Authorization::ConvertSidToStringSidW, GetTokenInformation, LookupAccountNameW,
+            TokenUser, SID, SID_NAME_USE, TOKEN_QUERY, TOKEN_USER,
+        },
+        System::{
+            Com::{
+                CoCreateInstance, CoInitializeEx, CoSetProxyBlanket, CoTaskMemFree,
+                CLSCTX_INPROC_SERVER, COINIT_MULTITHREADED, EOAC_NONE, RPC_C_AUTHN_LEVEL_CALL,
+                RPC_C_IMP_LEVEL_IMPERSONATE,
+            },
+            Rpc::{RPC_C_AUTHN_WINNT, RPC_C_AUTHZ_NONE},
+            Threading::{GetCurrentProcess, OpenProcessToken},
+            Wmi::{
+                IWbemLocator, IWbemServices, WbemLocator, WBEM_FLAG_CONNECT_USE_MAX_WAIT,
+                WBEM_FLAG_FORWARD_ONLY, WBEM_FLAG_RETURN_IMMEDIATELY, WBEM_INFINITE,
+            },
+        },
+        UI::Shell::{FOLDERID_Profile, SHGetKnownFolderPath, KNOWN_FOLDER_FLAG},
+    },
+};
 
+/// An identifier for a user.
+///
+/// This contains the text representation of the user's
+/// [SID](https://learn.microsoft.com/en-us/windows-server/identity/ad-ds/manage/understand-security-identifiers).
+/// It can be dereferenced to obtain this text.
+///
+/// It is currently an opaque type to prevent any potential unsoundness from having invalid SIDs
+/// passed to [`get_home_from_id`].
 #[repr(transparent)]
 #[derive(Debug)]
 pub struct UserIdentifier(String);
 
+/// This enumeration is the error type returned by this crate's functions
+/// on Windows.
 #[derive(Debug)]
 pub enum GetHomeError {
+    /// This represents an error as obtained from Windows' API.
     WindowsError(WinError),
-    FromUtf16Error(FromUtf16Error),
+    /// This represents an error when parsing UTF-16 text.
     Utf16Error(Utf16Error),
+    /// This represents an error when trying to represent a string that contains
+    /// a NUL byte `'\0'` as a C string.
     ContainsNul(ContainsNul<u16>),
+    /// This represents an error when a returned pointer was null when it was not expected to be
+    /// so.
     NullPointerResult,
-    WMINotBSTR,
 }
 
+/// This structure caches the results of the operations necessary to check the profile
+/// directory from an SID, see [`GetHomeInstance::get_home_from_id`]. This way, multiple
+/// queries can be performed at a smaller cost.
 pub struct GetHomeInstance(IWbemServices);
 
+/// This function will get the home directory of a user given their username. Internally,
+/// it calls [`get_id`] followed by [`get_home_from_id`].
+///
+/// Calling this function may present some issues if any other parts of the program use
+/// [`CoInitializeEx`](https://learn.microsoft.com/en-us/windows/win32/api/combaseapi/nf-combaseapi-coinitializeex).
+/// See [for Windows users](crate#for-windows-users) for more information.
 pub fn get_home<S: AsRef<str>>(username: S) -> Result<Option<PathBuf>, GetHomeError> {
     let Some(s) = get_id(username.as_ref())? else {
         return Ok(None);
@@ -32,16 +91,26 @@ pub fn get_home<S: AsRef<str>>(username: S) -> Result<Option<PathBuf>, GetHomeEr
     get_home_from_id(&s)
 }
 
+/// This function will get the home directory of a user given their identifier. This identifier can
+/// be obtained from [`get_id`]. The current process' user identifier can be found from
+/// [`get_my_id`]. Internally, this function calls [`GetHomeInstance::new`] followed by
+/// [`GetHomeInstance::get_home_from_id`].
+///
+/// Calling this function may present some issues if any other parts of the program use
+/// [`CoInitializeEx`](https://learn.microsoft.com/en-us/windows/win32/api/combaseapi/nf-combaseapi-coinitializeex).
+/// See [for Windows users](crate#for-windows-users) for more information.
 pub fn get_home_from_id(id: &UserIdentifier) -> Result<Option<PathBuf>, GetHomeError> {
     GetHomeInstance::new()?.get_home_from_id(id)
 }
 
+/// Get the identifier of a user given their username.
 pub fn get_id<S: AsRef<str>>(username: S) -> Result<Option<UserIdentifier>, GetHomeError> {
     unsafe {
         let username = U16CString::from_str(username)?;
         let mut sid_size = 0;
         let mut domain_size = 0;
         let mut peuse = SID_NAME_USE(0);
+        // get buffer length necessary for SID.
         if let Err(e) = LookupAccountNameW(
             None,
             PCWSTR(username.as_ptr()),
@@ -49,7 +118,7 @@ pub fn get_id<S: AsRef<str>>(username: S) -> Result<Option<UserIdentifier>, GetH
             &mut sid_size,
             PWSTR::null(),
             &mut domain_size,
-            &mut peuse
+            &mut peuse,
         ) {
             if e == ERROR_NONE_MAPPED.into() {
                 return Ok(None);
@@ -59,6 +128,8 @@ pub fn get_id<S: AsRef<str>>(username: S) -> Result<Option<UserIdentifier>, GetH
         }
         let layout = Layout::from_size_align(sid_size as usize, align_of::<SID>()).unwrap();
         let sid_buf = alloc_zeroed(layout);
+        // the domain is unfortunately necessary, otherwise the function will not operate
+        // correctly.
         let mut domain = vec![0; domain_size as usize];
         let psid = PSID(sid_buf.cast());
         LookupAccountNameW(
@@ -76,18 +147,15 @@ pub fn get_id<S: AsRef<str>>(username: S) -> Result<Option<UserIdentifier>, GetH
     }
 }
 
+/// Get the identifier of this process' user.
 pub fn get_my_id() -> Result<UserIdentifier, GetHomeError> {
     unsafe {
         // get the handle of the current process.
         let handle = GetCurrentProcess();
         let mut token_handle = HANDLE(0);
         // get a token to query information about the current process. this handle must be dropped
-        // manually with CloseHandle.
+        // manually with CloseHandle, as seen below.
         OpenProcessToken(handle, TOKEN_QUERY, &mut token_handle)?;
-        // this structure wraps the token handle so it is automatically dropped.
-        // this could be written better if try blocks were implemented, and it will probably
-        // be modified once they are stabilized for long enough (if ever)
-        let token_handle = token_handle;
         let mut buffer_size = 0;
         // get the length of the buffer requried for this query.
         if let Err(e) = GetTokenInformation(token_handle, TokenUser, None, 0, &mut buffer_size) {
@@ -96,9 +164,16 @@ pub fn get_my_id() -> Result<UserIdentifier, GetHomeError> {
                 return Err(e.into());
             }
         }
-        let layout = Layout::from_size_align(buffer_size as usize, align_of::<TOKEN_USER>()).unwrap();
+        let layout =
+            Layout::from_size_align(buffer_size as usize, align_of::<TOKEN_USER>()).unwrap();
         let buf_ptr = alloc_zeroed(layout);
-        if let Err(e) = GetTokenInformation(token_handle, TokenUser, Some(buf_ptr.cast()), buffer_size, &mut buffer_size) {
+        if let Err(e) = GetTokenInformation(
+            token_handle,
+            TokenUser,
+            Some(buf_ptr.cast()),
+            buffer_size,
+            &mut buffer_size,
+        ) {
             let _ = CloseHandle(token_handle);
             return Err(e.into());
         }
@@ -108,6 +183,7 @@ pub fn get_my_id() -> Result<UserIdentifier, GetHomeError> {
     }
 }
 
+/// Get the home directory of the current process' user.
 pub fn get_my_home() -> Result<Option<PathBuf>, GetHomeError> {
     unsafe {
         let out = SHGetKnownFolderPath(&FOLDERID_Profile, KNOWN_FOLDER_FLAG(0), None)?.0;
@@ -131,8 +207,9 @@ unsafe fn sid_to_string(sid: PSID) -> Result<UserIdentifier, GetHomeError> {
             // we already have an error. I won't check for this one.
             LocalFree(HLOCAL(str_pointer.0.cast()));
             return Err(e.into());
-        },
-    }.to_owned();
+        }
+    }
+    .to_owned();
     if !LocalFree(HLOCAL(str_pointer.0.cast())).0.is_null() {
         Err(WinError::from_win32())?;
     }
@@ -140,6 +217,7 @@ unsafe fn sid_to_string(sid: PSID) -> Result<UserIdentifier, GetHomeError> {
 }
 
 impl GetHomeInstance {
+    /// Construct this structure. This connects to the Windows Management Instrumentation.
     pub fn new() -> Result<Self, GetHomeError> {
         unsafe {
             const NAMESPACE_PATH: &str = "ROOT\\CIMV2";
@@ -184,12 +262,23 @@ impl GetHomeInstance {
         }
     }
 
+    /// Get the home directory of a user given their identifier.
     pub fn get_home_from_id(&self, id: &UserIdentifier) -> Result<Option<PathBuf>, GetHomeError> {
         unsafe {
-            let query_enum = self.0.ExecQuery(&BSTR::from("WQL"), &BSTR::from(format!("SELECT LocalPath FROM Win32_UserProfile WHERE SID = '{}'", id.0)), WBEM_FLAG_FORWARD_ONLY | WBEM_FLAG_RETURN_IMMEDIATELY, None)?;
+            let query_enum = self.0.ExecQuery(
+                &BSTR::from("WQL"),
+                &BSTR::from(format!(
+                    "SELECT LocalPath FROM Win32_UserProfile WHERE SID = '{}'",
+                    id.0
+                )),
+                WBEM_FLAG_FORWARD_ONLY | WBEM_FLAG_RETURN_IMMEDIATELY,
+                None,
+            )?;
             let mut ret = [None; 1];
             let mut ret_count = 0;
-            query_enum.Next(WBEM_INFINITE, &mut ret, &mut ret_count).ok()?;
+            query_enum
+                .Next(WBEM_INFINITE, &mut ret, &mut ret_count)
+                .ok()?;
             if ret_count == 0 {
                 return Ok(None);
             }
@@ -198,15 +287,11 @@ impl GetHomeInstance {
             let name = w!("LocalPath");
             let mut variant = VARIANT::default();
             let mut vt_type = 0;
-            ret.Get(
-                name,
-                0,
-                &mut variant,
-                Some(&mut vt_type),
-                None
-            )?;
+            ret.Get(name, 0, &mut variant, Some(&mut vt_type), None)?;
             let bstr = BSTR::try_from(&variant)?;
-            Ok(Some(U16Str::from_slice(bstr.as_wide()).to_os_string().into()))
+            Ok(Some(
+                U16Str::from_slice(bstr.as_wide()).to_os_string().into(),
+            ))
         }
     }
 }
@@ -229,9 +314,25 @@ impl From<ContainsNul<u16>> for GetHomeError {
     }
 }
 
-impl From<FromUtf16Error> for GetHomeError {
-    fn from(value: FromUtf16Error) -> Self {
-        Self::FromUtf16Error(value)
+impl fmt::Display for GetHomeError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::WindowsError(e) => write!(f, "windows error: {e}"),
+            Self::Utf16Error(e) => write!(f, "utf-16 error: {e}"),
+            Self::ContainsNul(e) => write!(f, "str contains NUL: {e}"),
+            Self::NullPointerResult => write!(f, "unexpected null pointer result"),
+        }
+    }
+}
+
+impl std::error::Error for GetHomeError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        match self {
+            Self::WindowsError(e) => Some(e),
+            Self::Utf16Error(e) => Some(e),
+            Self::ContainsNul(e) => Some(e),
+            Self::NullPointerResult => None,
+        }
     }
 }